Fixed numeric version too big if exceeds integer range

+++ src/main/groovy/com/apphance/ameba/ProjectConfiguration.groovy
@@ -18,7 +18,7 @@ class ProjectConfiguration  {
     File logDirectory
     File buildDirectory
     String versionString = "NOVERSION"
-    Integer versionCode = 0
+    Long versionCode = 0
 
     String bundleId
     String buildDate
